requirements.txt
torch
torchvision
gradio==3.50
gdown
Pillow
opencv-python
numpy
# 192.168.68.109